fix deletion issue

+++ src/main/java/com/bob/courses/CoursesRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 
 import java.util.List;
 @Repository
-public interface CoursesRepository extends JpaRepository<CoursesEntity,Long> {
+public interface CoursesRepository extends JpaRepository<CoursesEntity,Integer> {
     List<CoursesEntity> findByCourseId(int courseId);
 }
